Compare `cmsg_len` as `usize` for musl compatibility
`cmsg_len` is `usize` on glibc but `u32` on musl, so the comparison
against the required length fails to compile on musl without a cast.

+++ quinn-udp/src/linux.rs
@@ -87,7 +87,8 @@ impl LinuxError {
         let required =
             unsafe { libc::CMSG_LEN(size_of::<libc::sock_extended_err>() as _) as usize };
 
-        if cmsg.cmsg_len < required {
+        #[allow(clippy::unnecessary_cast)] // cmsg.cmsg_len defined as size_t
+        if (cmsg.cmsg_len as usize) < required {
             return None;
         }
